split RandomReadWriteSizeStream to separate file
